refactor and update glossary parsing
